move fetcher and api into folder
